add viz functions
